africa/requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
celery==5.3.1
click==8.1.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
flower==2.0.0
humanize==4.7.0
kombu==5.3.1
numpy==1.25.1
pika==1.3.2
prometheus-client==0.17.1
prompt-toolkit==3.0.39
python-dateutil==2.8.2
pytz==2023.3
scipy==1.11.1
six==1.16.0
sqlparse==0.4.4
tornado==6.3.2
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
